mysite/requirements.txt
# Dev environment
pip
autopep8

# App
asgiref==3.11.0
Django==5.2.9
sqlparse==0.4.4
typing_extensions==4.9.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
PyJWT==2.10.1
django-filter==25.2
stripe==9.1.0
